move test users call
